unexpectedError and tests, crud-only local fs
