+++ src/api.ts
@@ -36,36 +36,40 @@ export function isDir(node?: FileSystemNode | null): node is Directory {
     return node.type === 'dir';
 }
 
-export interface FileCreatedEvent{
+export interface UnexpectedErrorEvent {
+    type:'unexpectedError'
+}
+
+export interface FileCreatedEvent {
     type:'fileCreated',
     fullPath:string,
     newContent: string
 }
 
-export interface FileChangedEvent{
+export interface FileChangedEvent {
     type:'fileChanged',
     fullPath:string,
     newContent: string
 }
 
-export interface FileDeletedEvent{
+export interface FileDeletedEvent {
     type:'fileDeleted',
     fullPath:string
 }
 
-export interface DirectoryCreatedEvent{
+export interface DirectoryCreatedEvent {
     type:'directoryCreated',
     fullPath:string,
 }
 
-export interface DirectoryDeletedEvent{
+export interface DirectoryDeletedEvent {
     type:'directoryDeleted',
     fullPath:string
 }
 
 export type ListenerFn<T> = (event: T) => void;
 
-export interface EventAspect<S,O>{
+export interface EventAspect<S,O> {
     listeners(event: S, exists: boolean): Array<ListenerFn<O>> | boolean;
     listeners(event: S): Array<ListenerFn<O>>;
     on(event: S, fn: ListenerFn<O>, context?: any): this;
@@ -76,14 +80,15 @@ export interface EventAspect<S,O>{
     off(event: S, fn?: ListenerFn<O>, context?: any, once?: boolean): this;
 }
 
-export type FileSystemEventName = 'fileCreated' | 'fileChanged' | 'fileDeleted' | 'directoryCreated' | 'directoryDeleted';
-export const fileSystemEventNames: FileSystemEventName[] = ['fileCreated', 'fileChanged', 'fileDeleted', 'directoryCreated', 'directoryDeleted'];
-export type FileSystemEventHandler = FileCreatedEvent | FileChangedEvent | FileDeletedEvent | DirectoryCreatedEvent | DirectoryDeletedEvent;
+export type FileSystemEventName = 'unexpectedError' | 'fileCreated' | 'fileChanged' | 'fileDeleted' | 'directoryCreated' | 'directoryDeleted';
+export const fileSystemEventNames: FileSystemEventName[] = ['unexpectedError', 'fileCreated', 'fileChanged', 'fileDeleted', 'directoryCreated', 'directoryDeleted'];
+export type FileSystemEventHandler = UnexpectedErrorEvent | FileCreatedEvent | FileChangedEvent | FileDeletedEvent | DirectoryCreatedEvent | DirectoryDeletedEvent;
 
 export const fileSystemMethods = ['saveFile', 'deleteFile', 'deleteDirectory', 'loadTextFile', 'loadDirectoryTree', 'ensureDirectory'];
 
 export type EventEmitter =
-    EventAspect<'fileCreated', FileCreatedEvent> &
+    EventAspect<'unexpectedError', UnexpectedErrorEvent> &
+        EventAspect<'fileCreated', FileCreatedEvent> &
         EventAspect<'fileChanged', FileChangedEvent> &
         EventAspect<'fileDeleted', FileDeletedEvent> &
         EventAspect<'directoryCreated', DirectoryCreatedEvent> &
